agent-server/pyproject.toml
[project]
name = "agent-inference"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.4.8",
    "mlflow[genai]>=3.6.0",
    "anthropic>=0.79.0",
    "python-dotenv>=1.2.1",
    "psycopg2>=2.9.11",
    "opentelemetry-exporter-otlp>=1.39.1",
    "openinference-instrumentation-agno>=0.1.27",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "llama-index-core>=0.14.13",
    "llama-index-embeddings-fastembed>=0.5.0",
    "fastembed>=0.7.4",
    "qdrant-client>=1.16.2"
]
